backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
langchain
langchain-community
langchain-ollama
sentence-transformers